move getDisplayTime() to lib module
