+++ client/components/timer.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import getDisplayTime from '../lib/get-display-time';
 
 export default class Timer extends React.Component {
   constructor(props) {
@@ -37,32 +38,8 @@ export default class Timer extends React.Component {
     });
   }
 
-  getDisplayTime() {
-    let time = this.state.elapsed;
-    const min = Math.floor(time / 60000);
-    time = time - (min * 60000);
-    const sec = Math.floor(time / 1000).toString(10);
-    time = time - (sec * 1000);
-    const hundreths = Math.floor(time / 10).toString(10);
-    let displayedSec;
-    if (sec.length === 1) {
-      displayedSec = '0' + sec;
-    } else {
-      displayedSec = sec;
-    }
-    let displayedHundreths;
-    if (hundreths.length === 1) {
-      displayedHundreths = '0' + hundreths;
-    } else {
-      displayedHundreths = hundreths;
-    }
-    const displayedTime = `${min}:${displayedSec}.${displayedHundreths}`;
-
-    return displayedTime;
-  }
-
   render() {
-    const displayedTime = this.getDisplayTime();
+    const displayedTime = getDisplayTime(this.state.elapsed);
 
     const fullTimer = (
       <div
